+++ tests/acceptance/LeagueContainerAdapterTest.php
@@ -6,8 +6,6 @@ use League\Container\Container;
 
 final class LeagueContainerAdapterTest extends AbstractContainerAdapterTest
 {
-    use SupportsInflectorConfig;
-
     protected function setUp()
     {
         $this->container = new Container();