+++ tests/acceptance/SupportsInflectorConfig.php
@@ -3,12 +3,13 @@
 namespace tests\acceptance;
 
 use tests\mocks\ExampleClass;
+use tests\mocks\ExampleFactory;
 use tests\mocks\ExampleInterface;
 use TomPHP\ContainerConfigurator\Configurator;
 
 trait SupportsInflectorConfig
 {
-    public function testItSetsUpAnInflector()
+    public function testItSetsUpAnInflectorForClassNameFactory()
     {
         $config = [
             'di' => [
@@ -20,6 +21,65 @@ trait SupportsInflectorConfig
                 'inflectors' => [
                     ExampleInterface::class => [
                         'setValue' => ['test_value'],
+                    ],
+                ],
+            ],
+        ];
+
+        Configurator::apply()
+            ->configFromArray($config)
+            ->to($this->container);
+
+        assertEquals(
+            'test_value',
+            $this->container->get('example')->getValue()
+        );
+    }
+
+    public function testItSetsUpAnInflectorForServiceFactory()
+    {
+        $config = [
+            'class_name' => ExampleClass::class,
+            'di'         => [
+                'services' => [
+                    'example' => [
+                        'factory'   => ExampleFactory::class,
+                        'arguments' => [
+                            'config.class_name',
+                        ],
+                    ],
+                ],
+                'inflectors' => [
+                    ExampleInterface::class => [
+                        'setValue' => ['test_value'],
+                    ],
+                ],
+            ],
+        ];
+
+        Configurator::apply()
+            ->configFromArray($config)
+            ->to($this->container);
+
+        assertEquals(
+            'test_value',
+            $this->container->get('example')->getValue()
+        );
+    }
+
+    public function testItResolvesInflectorArguments()
+    {
+        $config = [
+            'argument' => 'test_value',
+            'di'       => [
+                'services' => [
+                    'example' => [
+                        'class' => ExampleClass::class,
+                    ],
+                ],
+                'inflectors' => [
+                    ExampleInterface::class => [
+                        'setValue' => ['config.argument'],
                     ],
                 ],
             ],