+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - TDB
+### Added
+  * Inflector support for Pimple.
+
 ## 0.5.2 - 2016-09-30
 ### Added
   * Service aliases.